pyproject.toml
[tool.poetry]
name = "mango-fzco-test"
version = "0.1.0"
description = ""
authors = ["Aleksei Chistov <aleksey.chistov@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.12"
fastapi-cli = "^0.0.7"
websockets = "^15.0.1"
python-multipart = "^0.0.20"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.40"}
asyncpg = "^0.30.0"
pydantic-settings = "^2.8.1"
pyjwt = "^2.10.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
ruff = "^0.11.2"
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 90

[tool.isort]
profile = "black"

[tool.ruff]
[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "ANN",
    "B008",
    "B904",
    "B011",
    "BLE001",
    "C401",
    "C901",
    "COM",
    "D",
    "E501",
    "E711",
    "E731",
    "ERA001",
    "FA102",
    "FAST002",
    "FBT003",
    "FIX002",
    "G001",
    "G004",
    "LOG002",
    "N806",
    "PERF401",
    "PLR0912",
    "PLR0913",
    "PLR0915",
    "PLR1714",
    "PLR1722",
    "PLR2004",
    "PLR5501",
    "PLW0603",
    "PTH123",
    "RUF003",
    "RUF012",
    "S",
    "SIM102",
    "SIM103",
    "SIM108",
    "SIM117",
    "T201",
    "TD",
    "TRY300",
    "UP032",
    "UP035",
]

[tool.pytest.ini_options]
addopts = ["-ra", "-q"]
pythonpath = "src"
testpaths = ["src/test.py"]
asyncio_mode = "auto"
